Add files via upload
In this project we are using SQL Server. This will include the highest revenue generators, analyzing sales trends by region, and examining the 2022 and 2023 monthly changes. It will also consider the respective product category sales and months of sale while comparing year-on-year subcategory profit growth. This is a very detailed analysis helping bring out identified areas for strategic improvement and informed decision-making across the business.